+++ Gruntfile.js
@@ -53,6 +53,7 @@ module.exports = function (grunt) {
         '<%= nxt_dir.vendor %>/leaflet-dist/leaflet.js',
         '<%= nxt_dir.vendor %>/angular-ui-bootstrap/ui-bootstrap-tpls-0.10.0.min.js',
         '<%= nxt_dir.vendor %>/lodash/dist/lodash.min.js',
+        '<%= nxt_dir.vendor %>/raven-js/dist/raven.min.js'
       ],
       testfiles: [
         '<%= nxt_dir.vendor %>/angular-mocks/angular-mocks.js',
@@ -93,7 +94,7 @@ module.exports = function (grunt) {
             '<%= appfiles %>',
             '<%= nxt_dir.test %>/**/*.js'
           ],
-          tasks: ['test']
+          tasks: ['test', 'jsdoc']
         },
         styles: {
           files: ['<%= nxt_dir.src %>/assets/css/{,*/}*.css'],
@@ -133,6 +134,15 @@ module.exports = function (grunt) {
               '<%= nxt_dir.vendor %>/ng-table/ng-table.min.css',
               '<%= nxt_dir.vendor %>/lizard-iconfont/lizard/dest/css/Lizard.css'
             ]
+          }
+        }
+      },
+      // produces docs
+      jsdoc: {
+        dist: {
+          src: ['<%= appfiles %>'],
+          options: {
+            destination: 'doc'
           }
         }
       },
@@ -246,6 +256,7 @@ module.exports = function (grunt) {
     'cssmin',
     'html2js',
     'concat',
+    'jsdoc',
     'copy:dist'
   ]);
   grunt.registerTask('plakhetaanelkaar', [