+++ source/app/lizard-nxt.js
@@ -11,8 +11,33 @@ var app = angular.module("lizard-nxt", [
   'ngCsv',
   'ui.bootstrap',
   'ui.utils',
-  'ngTable'
 ]);
+
+/**
+ * Setup Raven if available.
+ * Raven is responsible for logging to https://sentry.lizard.net
+ */
+if (window.Raven) {
+  Raven.config('https://ceb01dd84c6941c8aa20e16f83bdb55e@sentry.lizard.net/19',
+  {
+    // limits logging to staging and prd
+    whitelistUrls: [/nxt\.lizard\.net/, /staging\.lizard\.net/]
+  }).install();
+}
+
+/**
+ * Decorator for ngExceptionHandler to log exceptions to sentry
+ */
+app.config(function ($provide) {
+  $provide.decorator("$exceptionHandler", function ($delegate) {
+      return function (exception, cause) {
+          $delegate(exception, cause);
+          Raven.captureException(exception, {
+            extra: {cause: cause}
+          });
+        };
+    });
+});
 
 /**
  * Change default angular tags to prevent collision with Django tags.
@@ -31,61 +56,76 @@ app.config(function ($locationProvider) {
 });
 
 /**
- * Master controller
  *
- * Overview
- * ========
+ * @name MasterController
+ * @class MasterCtrl
+ * @memberOf app
+ * @requires UtilService
+ * @requires EventService
+ * @requires CabinetService
+ * @requires RasterService
+ *
+ * @summary Master controller
+ *
+ * @description
+ * ## Overview
  *
  * Defines general models and gets data from server; functions that are not
  * relevant for rootscope live in their own controller
  *
  * Directives watch models in MasterCtrl and respond to changes in those models
  * for example, a user zooms in on the timeline, the timeline directive sets
- * the temporal.extent on the state.temporal; a map directive watches state.temporal
- * and updates map objects accordingly.
+ * the temporal.extent on the state.temporal; a map directive watches
+ * state.temporal and updates map objects accordingly.
  *
- * Models
- * ======
+ * ## Models
  *
- * Application state
- * -----------------
+ * Application state >> link to state property in this module.
+ *
  * state.mapState => spatial state
  * state.timeState => temporal state
  * state.tools => active tool(s)
  * user.profile
  *
- * Data
- * ----
+ * Data >> link to data properties in this module.
  * data.active
  * data.objects
  * data.events
  * data.timeseries
  * data.aggregates
  *
- * TODO / Refactor
- * ---------------
+ * ## TODO / Refactor
  *
  * Stuff to reconsider, rethink, refactor:
  *
- * * [ ] Refactor map controller and directives
- * * [-] Refactor master controller (states, data!)
- * * [+] Refactor timeline out of mapState with its own scope
- * * [+] Refactor index.html and base-debug.html
- * * [ ] Fix + document Gruntfile.js / workflow
- * * [ ] Refactor css (csslint, -moz and -webkit)
- * * [ ] Move or delete common directory in source
- * * [+] Refactor timeline controller and directive
- * * [ ] Move event logic to event controller (on event / layer tag)
- * * [+] Move animation logic to animation controller (on timeline tag)
+ * - [ ] Refactor map controller and directives
+ * - [-] Refactor master controller (states, data!)
+ * - [+] Refactor timeline out of mapState with its own scope
+ * - [+] Refactor index.html and base-debug.html
+ * - [ ] Fix + document Gruntfile.js / workflow
+ * - [ ] Refactor css (csslint, -moz and -webkit)
+ * - [ ] Move or delete common directory in source
+ * - [+] Refactor timeline controller and directive
+ * - [ ] Move event logic to event controller (on event / layer tag)
+ * - [+] Move animation logic to animation controller (on timeline tag)
 
  */
 app.controller("MasterCtrl",
   ["$scope", "$http", "$q", "$filter", "$compile", "CabinetService", "RasterService",
-   "UtilService", "EventService", "TimeseriesService", "ngTableParams",
+   "UtilService", "EventService", "TimeseriesService",
   function ($scope, $http, $q, $filter, $compile, CabinetService, RasterService,
-            UtilService, EventService, TimeseriesService, ngTableParams) {
+            UtilService, EventService, TimeseriesService) {
 
   // BOX MODEL
+  /**
+   * @memberOf app.MasterCtrl
+   * @summary Box model
+   *
+   * @description Box model holds properties to render the omnibox.
+   *
+   * @property {object} box - Box model
+   * @property {boolean} box.detailMode - Detail mode, defaults to false.
+   */
   $scope.box = {
     contextSwitchMode: false, // Switch between card or fullscreen
     query: null, // Search bar query
@@ -119,22 +159,37 @@ app.controller("MasterCtrl",
     'pumpstation_sewerage': 'Rioolgemaal'
   };
 
+  /**
+   * @function
+   * @memberOf app.MasterCtrl
+   *
+   * @summary Get translation for text
+   * @desc Get translation for text
+   *
+   * @param {string} text - Text to translate.
+   * @returns {string} text - Translated text.
+   */
   $scope.gettext = function (text) {
     return $scope.translations[text];
   };
 
 
   /**
-   * Toggle tool from "name" to "none"
+   * @function
+   * @memberOf app.MasterCtrl
    *
-   * Sets tool.active model on scope to name of the tool if tool disabled
+   * @summary Toggle tool from "name" to "none".
+   *
+   * @desc Sets tool.active model on scope to name of the tool if tool disabled
    * or "none" if tool is already enabled.
    *
    * @param {string} name name of the tool to toggle
    *
    */
   $scope.toggleTool = function (name) {
-
+    if (name === 'intersect') {
+      $scope.box.type  = 'intersect';
+    }
     if ($scope.tools.active === name) {
       $scope.tools.active = 'none';
       $scope.box.type = 'extentAggregate';
@@ -169,7 +224,10 @@ app.controller("MasterCtrl",
     baselayerChanged: Date.now(),
     enabled: false,
     bounds: null,
+    pixelCenter: null,
+    zoom: null,
     here: null, // Leaflet point object describing a users location of interest
+    userHere: null, // Geographical location of the users mouse
     geom_wkt: '',
     mapMoving: false
   };
@@ -180,15 +238,11 @@ app.controller("MasterCtrl",
 
   var now = Date.now();
   var day = 24 * 60 * 60 * 1000;
-  var tomorrow = now + day;
-  var twoDaysAgo = now - 2 * day;
-  var sevenDaysAgo = now - 7 * day;
-  var lastVisit = CabinetService.lastVisitUtime;
-  var start = Math.max(sevenDaysAgo, Math.min(twoDaysAgo, lastVisit)) || sevenDaysAgo;
+
   // TIME MODEL
   $scope.timeState = {
-    start: start,
-    end: tomorrow,
+    start: now - 2 * day,
+    end: now + day,
     changedZoom: Date.now(),
     zoomEnded: null,
     hidden: undefined,