+++ CHANGES.rst
@@ -4,13 +4,18 @@ Changelog of lizard-nxt client
 unreleased
 ----------
 
+- Add jsdoc-conf.json, configuration file for jsdoc.
+
+- Intersection tool shows generic functionality for all 3-or-more-d layers
 
 Release 0.1
 -----------
 
+- Layer chooser is now a directive and has a background-image.
+
 - All clicks on the map result all data available to that location
 
-- ExtentAggregat is the default card displaying an extent summary of 
+- ExtentAggregat is the default card displaying an extent summary of
   all active data layers
 
 - Layers are mentioned in slug of URL
@@ -91,6 +96,6 @@ Release 0.1
 
 - JSHint cleanup.
 
-- Gruntfile cleanup. 
+- Gruntfile cleanup.
 
 - Got tests to run.